Switch from m2r to sphinx-mdinclude
m2r is unmaintained [0]. It relies on mistune, but is not compatible with
mistune versions >= 2.0 [1]. This makes it impossible to update mistune,
even though there is a CVE in the package [2].

Switch to sphinx-mdinclude instead.

There is also m2r2, but it is not a suitable alternative because they
simply force the use of an older mistune version.

[0] https://github.com/sphinx-contrib/openapi/issues/123
[1] https://github.com/miyakogi/m2r/issues/66
[2] https://github.com/advisories/GHSA-fw3v-x4f2-v673
[3] https://github.com/CrossNox/m2r2/pull/43